# WIP: custom serialization for field set