--- a/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.proto
+++ b/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.proto
@@ -215,18 +215,7 @@ message ExportOptions {
 // Fields stores a set of fields in a data structure like a Trie.
 // To understand how this is used, see: https://github.com/kubernetes-sigs/structured-merge-diff
 message Fields {
-  // Map stores a set of fields in a data structure like a Trie.
-  //
-  // Each key is either a '.' representing the field itself, and will always map to an empty set,
-  // or a string representing a sub-field or item. The string will follow one of these four formats:
-  // 'f:<name>', where <name> is the name of a field in a struct, or key in a map
-  // 'v:<value>', where <value> is the exact json formatted value of a list item
-  // 'i:<index>', where <index> is position of a item in a list
-  // 'k:<keys>', where <keys> is a map of  a list item's key fields to their unique values
-  // If a key maps to an empty Fields value, the field that key represents is part of the set.
-  //
-  // The exact format is defined in k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal
-  map<string, Fields> map = 1;
+  optional k8s.io.apimachinery.pkg.runtime.RawExtension map = 2;
 }
 
 // GetOptions is the standard query options to the standard REST get call.