--- a/staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/fields_test.go
+++ b/staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/fields_test.go
@@ -74,7 +74,7 @@ func TestFieldsToSetError(t *testing.T) {
 					},
 				},
 			},
-			errString: "invalid character",
+			errString: "error found",
 		},
 	}
 
@@ -97,7 +97,7 @@ func TestSetToFieldsError(t *testing.T) {
 	}{
 		{
 			set:       *fieldpath.NewSet(invalidPath),
-			errString: "Invalid type of path element",
+			errString: "invalid PathElement",
 		},
 	}