… go.sum
… pkg/kubectl/generated/bindata.go
… staging/src/k8s.io/apiextensions-apiserver/go.sum
… staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.pb.go
… staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/zz_generated.deepcopy.go
… staging/src/k8s.io/apiserver/go.sum
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/BUILD
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/pathelement.go
… staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/pathelement_test.go
… staging/src/k8s.io/kube-aggregator/go.sum
… staging/src/k8s.io/legacy-cloud-providers/go.sum
… staging/src/k8s.io/sample-apiserver/go.sum
… vendor/modules.txt
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/BUILD
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/serialize-pe.go
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/serialize.go
… vendor/sigs.k8s.io/structured-merge-diff/fieldpath/set.go
… vendor/sigs.k8s.io/structured-merge-diff/value/BUILD
… vendor/sigs.k8s.io/structured-merge-diff/value/fastjson.go
… vendor/sigs.k8s.io/structured-merge-diff/value/value.go