This is to show how https://github.com/kubernetes-sigs/structured-merge-diff/pull/90 might be used.