@lavalamp 2019-07-23
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2065260       1903735       -7.82%
BenchmarkUpdateNewObject-12     2415312       2178154       -9.82%
BenchmarkRepeatedUpdate-12      967089        784110        -18.92%
BenchmarkSetToFields-12         22058         8862          -59.82%
BenchmarkFieldsToSet-12         15648         11641         -25.61%
BenchmarkYAMLToTyped-12         65407         63847         -2.39%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      7168           6861           -4.28%
BenchmarkUpdateNewObject-12     8396           7953           -5.28%
BenchmarkRepeatedUpdate-12      2887           2342           -18.88%
BenchmarkSetToFields-12         130            42             -67.69%
BenchmarkFieldsToSet-12         91             94             +3.30%
BenchmarkYAMLToTyped-12         152            152            +0.00%
```

I expected more impact so there's probably some more to do here.