@apelisse 2019-07-24
This is very important for the read path too, so I'm excited to get that merged soon.