@lavalamp 2019-07-24
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      2230683       1957508       -12.25%
BenchmarkUpdateNewObject-12     2624417       2166853       -17.43%
BenchmarkRepeatedUpdate-12      1031653       765925        -25.76%
BenchmarkSetToFields-12         24778         5217          -78.95%
BenchmarkFieldsToSet-12         33876         10303         -69.59%
BenchmarkYAMLToTyped-12         68377         62136         -9.13%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      8632           6827           -20.91%
BenchmarkUpdateNewObject-12     11210          7913           -29.41%
BenchmarkRepeatedUpdate-12      3057           2288           -25.16%
BenchmarkSetToFields-12         124            14             -88.71%
BenchmarkFieldsToSet-12         147            82             -44.22%
BenchmarkYAMLToTyped-12         150            152            +1.33%
```

This needs to fix up validation code and comments and other things before it's acceptable to merge.